added a cleanup func to cluster
instead of just exposing the internal array, a function is used instead.
